test: move the frame processor setup helper into tests/
`frame_processor_setup()` builds a FrameProcessorSetup with no pipeline worker
so a processor, controller or strategy can be driven on its own. That serves
Pipecat's own tests rather than anyone writing a bot, so it sits with the other
shared test helpers instead of in the shipped `pipecat.tests.utils`.
