+++ src/pipecat/tests/utils.py
@@ -10,7 +10,6 @@ import asyncio
 from collections.abc import Awaitable, Callable, Sequence
 from dataclasses import dataclass
 
-from pipecat.clocks.system_clock import SystemClock
 from pipecat.frames.frames import (
     EndFrame,
     Frame,
@@ -21,35 +20,8 @@ from pipecat.frames.frames import (
 from pipecat.observers.base_observer import BaseObserver, FramePushed
 from pipecat.pipeline.pipeline import Pipeline
 from pipecat.pipeline.worker import PipelineParams, PipelineWorker
-from pipecat.processors.frame_processor import (
-    FrameDirection,
-    FrameProcessor,
-    FrameProcessorSetup,
-)
-from pipecat.utils.asyncio.task_manager import BaseTaskManager
+from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
 from pipecat.workers.runner import WorkerRunner
-
-
-def frame_processor_setup(
-    task_manager: BaseTaskManager | None = None, **kwargs
-) -> FrameProcessorSetup:
-    """Build a setup configuration for a processor used outside a pipeline.
-
-    Tests that drive a single processor, controller or turn strategy need a
-    :class:`FrameProcessorSetup` without a surrounding :class:`PipelineWorker`.
-    Fields a test doesn't exercise can be left unset.
-
-    Args:
-        task_manager: The task manager the processor should run its tasks on.
-            Omit it when the processor under test creates no tasks.
-        **kwargs: Any :class:`FrameProcessorSetup` field to override.
-
-    Returns:
-        A setup carrying a system clock and no pipeline worker.
-    """
-    kwargs.setdefault("clock", SystemClock())
-    kwargs.setdefault("pipeline_worker", None)
-    return FrameProcessorSetup(task_manager=task_manager, **kwargs)  # pyright: ignore[reportArgumentType]
 
 
 @dataclass
